Skip schema object check
It's not easy to trace how/why, but a singleton/cached value of Activity
type in internal is colliding with the one we want to use in the Public
API. It's very difficult to trace how the activity is being swept in.

Found different types with the same name in the schema: Activity,
Activity.